+++ lib/jekyll-pandoc-exports.rb
@@ -1,2 +1,3 @@
 require "jekyll-pandoc-exports/version"
 require "jekyll-pandoc-exports/generator"
+require "jekyll-pandoc-exports/command"